Add focus styles for navmenu
Wrapped the navmenu icon in a button element and tweaked the styles so that
a keyboard can be used to open the nav drawer. This also adds more semantic
HTML, as an svg element should not be used as a button.

Added the same hover styles for drawer items to their focus state as well.